Ask AI to generate persona profiles
In this folder, I have created some files that should represent the personas of the different kind of people that will interact with the project. The project will be a system that can query data from a database, apply one or several data transformation based on rules provided by a functional team, and send that transformed data to the client.

The purpose of these files is to later use them as base to make sure the project respects the wish and desires of each persona, while ensuring to limits their fears.

Help me complete these personas, knowing that:

the client wants a simple system that will "just work". They are not intersted in the technical details. While it is a single persona, it can represent different clients that won't interact with one another.
the developer from the functional team wants to be able to develop a script or binary that will be used by the platform without knowing its details. It should be a black box that "just works", but they should have some confidence it terms of observability.
the developer from the platform team wants the system to work, to be able to make improvements and bug fixs, deliver new versions without impacting the functional developers or the existing clients.
the product owner is in contact with the client and the team of functional developers. It's objective is to offer to the client a solution that uses the platform and ask the function team to develop scripts or binaries that answer the client demans. They also want guarantee and observability on the overall platform.
the tester from the functional team wants to be able to test the scripts or binaries developed by the developers from the same team, but also test the integration into the platform.
the tester from the platform team needs to ensure that any evolution of the platform doesn't break existing workflows.
Each persona shoud be organized in a similar way with at least:

a short description of what the persona is
a section about the needs
a section about the wishes
a section about the fears or blockers