Ask AI to summarize the purpose of the personas
Let's now add a readme to that folder that explains the purpose of these persona documents.

It should point out that we will use these personas throughout the development of the project to ensure that:
- each needs of each persona MUST be met
- fears from any persona CAN'T be realized
- some wishes SHOULD be possible

Make sure the readme is clear, short and concise, and can be understood unambiguously by a human or an AI agent.