Ask AI to create a proposal about the transformation format
Let's explore the first open question about "transformation format". We actually already know that the functional team is more familiar with .NET and C#. What we don't know yet is whether the deliverables they will produce will be:
- simple scripts written in C# (but the compilation will be performed in the platform)
- binaries (e.g. nuget packages) that will respect some contract interfaces from a SDK made by the platform team
- containerized applications that will expose well-known REST endpoints that the platform will know how to call

Write a draft proposal for these options in the 1_proposals folder. On top of exploring them, it should take into account industry best practices as well as list existing (ideally open-source) libraries or frameworks that solve this issue.

For the purpose of this task, it is ok to do web searches to gather relevant information, or look for extisting technologies.