Ask AI to create a proposal about the CDC
Now let's explore the Change Data Capture (CDC) Mechanism. Member of our teams have mentioned Debezium as a possible candidate. We want to look for other alternatives that are:
- open-source
- developed with or related to .NET stack
- cloud-agnostic (we are excluding from our search Azure-specific technologies, though they can be mentioned as references)
- have good reputation within the industry or possess comprehensive documentation and tutorials
- can work with different database vendors

Write a draft proposal for these options in the 1_proposals folder.

For the purpose of this task, it is ok to do web searches to gather relevant information, or look for extisting technologies.