Ask AI to fix a confusion about the location of the source databases
Thanks. There is a slight confusion about the architecture. The ETL platform will not stream data from client's database. The source databases are actually in our premise. They are currently part of applications that we provide to the client and that they use (and hence contain data specific to each client).

Although it doesn't change much in term of arcitectural decisions, we should avoid that confusion and maybe generalize to any database regardless of whether they are hosted by us, are from the same client asking for this ETL or from any other third-party.

Consider correcting the root README, the architecture overview document, and any other document that was having this confusion. Don't change their content too much, it should be matter of removing or replacing a few words or sentences.